snowpark_project/requirements.txt
snowflake-snowpark-python
python-dotenv
pytest
